* feat: add support for generating templates for multiple days
